+++ assets/javascripts/kitten/components/carousel/carousel/carousel.js
@@ -64,6 +64,7 @@ const propTypesPositions = PropTypes.oneOf(['top', 'right', 'bottom', 'left'])
 class CarouselBase extends Component {
   static defaultProps = {
     hidePaginationOnMobile: false,
+    showOtherPages: false,
     paginationPosition: {
       default: 'right',
       fromM: 'bottom',
@@ -77,6 +78,7 @@ class CarouselBase extends Component {
     viewportIsMOrLess: PropTypes.bool.isRequired,
     viewportIsXSOrLess: PropTypes.bool.isRequired,
     hidePaginationOnMobile: PropTypes.bool,
+    showOtherPages: PropTypes.bool,
     paginationPosition: PropTypes.shape({
       default: propTypesPositions,
       fromXxs: propTypesPositions,
@@ -161,6 +163,7 @@ class CarouselBase extends Component {
       baseItemMarginBetween,
       viewportIsXSOrLess,
       viewportIsMOrLess,
+      showOtherPages,
     } = this.props
     const { indexPageVisible, numColumns, numPages } = this.state
     const itemMarginBetween = getMarginBetweenAccordingToViewport(
@@ -179,6 +182,7 @@ class CarouselBase extends Component {
         itemMarginBetween={itemMarginBetween}
         onResizeInner={this.onResizeInner}
         goToPage={this.goToPage}
+        showOtherPages={showOtherPages}
       />
     )
   }