+++ CHANGELOG.md
@@ -4,8 +4,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Add `showOtherPages` and `pagesClassName` to `Carousel`.
-- Fix: Remove `isActive` styles to `EngagementCard`.
+## [2.58.0] - 2020-03-11
+
+Feature:
+- Add `showOtherPages` and `pagesClassName` to `Carousel`.
+
+Fix:
+- Remove `isActive` styles to `EngagementCard`.
 
 ## [2.57.0] - 2020-03-06